+++ _Typings/utils/JSONExtension.d.ts
@@ -1,19 +1,9 @@
 // Licensed under the Apache License: http://www.apache.org/licenses/LICENSE-2.0
 // For details: https://github.com/Withered-Flower-0422/BST/blob/main/LICENSE
 
-export {}
+import { AddTypeTo, TurnTypeTo, PickStringKey } from "utils"
 
 declare const jsonDataType: unique symbol
-
-type AddNullToNumber<T> = T extends number
-    ? T | null
-    : T extends object
-      ? { [K in keyof T]: AddNullToNumber<T[K]> }
-      : T
-
-type StringifyReturnType<T> = T extends Function | undefined | symbol
-    ? undefined
-    : json<AddNullToNumber<T>> // JSON.stringify() returns `null` for `NaN` and `Infinity`.
 
 declare global {
     /** A valid JSON string. */
@@ -26,6 +16,14 @@ declare global {
             value: T extends bigint ? never : T,
             replacer?: null,
             space?: number,
-        ): StringifyReturnType<T>
+        ): json<
+            PickStringKey<
+                AddTypeTo<
+                    TurnTypeTo<T, undefined, Function | symbol>,
+                    null,
+                    number
+                >
+            >
+        >
     }
 }